\name{doublysketchingglms-package}
\alias{doublysketchingglms-package}
\alias{doublysketchingglms}
\docType{package}
\title{
\packageTitle{doublysketchingglms}
}
\description{
\packageDescription{doublysketchingglms}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{doublysketchingglms}
\packageIndices{doublysketchingglms}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{doublysketchingglms}

Maintainer: \packageMaintainer{doublysketchingglms}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS ~~
~~ in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
