% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Gamma_Inverse_IRLS}
\alias{Gamma_Inverse_IRLS}
\title{Fit Gamma-Inverse regression using IRLS in C++}
\usage{
Gamma_Inverse_IRLS(X, y, beta, iters = 10L, backtrack = 100L, eps = 1e-10)
}
\arguments{
\item{X}{A numeric matrix with the desired covariates. If an intercept is
desired, then a column of ones should be part of this input matrix.}

\item{y}{A vector of responses taking values zero or one. Should be the same
length as the number of rows of X.}

\item{beta}{The initial value of the coefficients.}

\item{iters}{The maximum number of IRLS iterations to run.}

\item{backtrack}{An integer determining the number of times to backtrack an
invalid parameter update. Each backtracking step consists of halving the
update difference, which is as done by \link[stats:glm]{stats::glm.fit} internally.}

\item{eps}{The convergence threshold for the L1 norm between successive
iterations of the parameters during the IRLS iteration.}
}
\value{
Returns a list with three elements: \code{B} which is a matrix where each row is
the parameter at the corresponding IRLS iteration, \code{cov} is the
Hessian matrix at the final iteration, and \code{dispersion} which is the fitted
dispersion value.
}
\description{
Fit Gamma-Inverse regression using IRLS in C++
}
