% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Gamma_Inverse_wIRLS}
\alias{Gamma_Inverse_wIRLS}
\title{Fit a weighted Gamma-Inverse regression using IRLS in C++}
\usage{
Gamma_Inverse_wIRLS(X, y, w, beta, iters = 10L, backtrack = 100L, eps = 1e-10)
}
\arguments{
\item{X, y, beta, iters, backtrack, eps}{As in \link{Gamma_Inverse_IRLS}.}

\item{w}{A vector of weights.}
}
\value{
Returns a list with three elements: \code{B} which is a matrix where each row is
the parameter at the corresponding IRLS iteration, \code{cov} is the
Hessian matrix at the final iteration, and \code{dispersion} which is the fitted
dispersion value.
}
\description{
This is a helper function to perform weighted Gamma-Inverse GLM fit for the
Optimal Distributed Subsampling for Maximum Quasi-Likelihood Estimators With
Massive Data method by Yu et al. (2022). This is so that the code is similar
and so the speeds are comparable.
}
