% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Logistic_wIRLS}
\alias{Logistic_wIRLS}
\title{Fit a weighted logistic regression using IRLS in C++}
\usage{
Logistic_wIRLS(X, y, w, beta, iters = 10L, eps = 1e-10)
}
\arguments{
\item{X, y, beta, iters, eps}{As in \link{Logistic_IRLS}.}

\item{w}{A vector of weights.}
}
\value{
Returns a list with two elements: \code{B} which is a matrix where each row is
the parameter at the corresponding IRLS iteration and \code{cov} is the
Hessian matrix at the final iteration.
}
\description{
This is a helper function to perform weighted logistic GLM fit for the
Optimal Distributed Subsampling for Maximum Quasi-Likelihood Estimators With
Massive Data method by Yu et al. (2022). This is so that the code is similar
and so the speeds are comparable.
}
