% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Poisson_IRLS}
\alias{Poisson_IRLS}
\title{Fit Poisson-Log regression using IRLS in C++}
\usage{
Poisson_IRLS(X, y, beta, iters = 10L, eps = 1e-10)
}
\arguments{
\item{X}{A numeric matrix with the desired covariates. If an intercept is
desired, then a column of ones should be part of this input matrix.}

\item{y}{A vector of responses taking values zero or one. Should be the same
length as the number of rows of X.}

\item{beta}{The initial value of the coefficients.}

\item{iters}{The maximum number of IRLS iterations to run.}

\item{eps}{The convergence threshold for the L1 norm between successive
iterations of the parameters during the IRLS iteration.}
}
\value{
Returns a list with two elements: \code{B} which is a matrix where each row is
the parameter at the corresponding IRLS iteration and \code{cov} is the
Hessian matrix at the final iteration.
}
\description{
Fit Poisson-Log regression using IRLS in C++
}
\details{
This function skips a lot of sanity checks and input validation that the
usual R functions do in the name of speed. There are also no guardrails
against unstable numerical computations.
}
